Complete admin dashboard project.
While the layout is there, I did not tie the
buttons to any sort of functionality and
omitted some list items from being links. The site
while decently flexible, is not perfectly dynamic
for all screen sizes. Would fix that as well, but
other than that, happy with the replication from
the original.